LayoutTests/imported/w3c/web-platform-tests/html/browsers/the-window-object/window-open-noopener-expected.txt

CONSOLE MESSAGE: line 1: TypeError: this.checkValues is not a function. (In 'this.checkValues(w2)', 'this.checkValues' is undefined)
CONSOLE MESSAGE: line 1: TypeError: this.checkValues is not a function. (In 'this.checkValues(w2)', 'this.checkValues' is undefined)
CONSOLE MESSAGE: line 1: TypeError: this.checkValues is not a function. (In 'this.checkValues(w2)', 'this.checkValues' is undefined)

Harness Error (FAIL), message = Test named 'window.open() with 'noopener' should not reuse existing target' specified 1 'cleanup' function, and 1 failed.

FAIL window.open() with 'noopener' should not reuse existing target Can't find variable: BroadcastChannel
NOTRUN noopener needs to be present as a token on its own 
NOTRUN noopener needs to be present as a token on its own again 
NOTRUN noopener needs to be present as a token on its own yet again 
NOTRUN Trailing noopener should work 
NOTRUN Leading noopener should work 
NOTRUN Interior noopener should work 
NOTRUN noopener window.open targeting _self 
NOTRUN noopener window.open targeting _parent 
NOTRUN noopener window.open targeting _top 
